requirements.txt
Django==3.2.16
Pillow==9.3.0
pytz==2023.3.post1
sqlparse==0.4.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter
